Update supplementary scripts

+++ linux-image.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 set -x
-RAM=1
-IMAGE_FOLDER=""
+export PATH=$PWD/gcc-linaro-6.3.1-2017.05-x86_64_aarch64-linux-gnu/bin:$PATH
+#RAM=1
+RAM=0
+#PROXY="http://127.0.0.1:3142"
+PROXY=""
+IMAGE_FOLDER="img/"
 IMAGE_VERSION="linux-libretech"
 IMAGE_DEVICE_TREE="amlogic/meson-gxl-s905x-libretech-cc"
 if [ ! -z "$1" ]; then
@@ -26,7 +30,7 @@ mkdir -p "$IMAGE_FOLDER"
 if [ $RAM -ne 0 ]; then
 	mount -t tmpfs -o size=1G tmpfs $IMAGE_FOLDER
 fi
-truncate -s 1G "${IMAGE_FOLDER}${IMAGE_FILE_NAME}"
+truncate -s 4G "${IMAGE_FOLDER}${IMAGE_FILE_NAME}"
 fdisk "${IMAGE_FOLDER}${IMAGE_FILE_NAME}" <<EOF
 o
 n
@@ -67,20 +71,58 @@ cp ${IMAGE_VERSION}/arch/arm64/boot/dts/$IMAGE_DEVICE_TREE.dtb p1/${IMAGE_DEVICE
 PATH=$PWD/gcc/bin:$PATH make -C ${IMAGE_VERSION} ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- headers_install INSTALL_HDR_PATH=$PWD/p2/usr/
 PATH=$PWD/gcc/bin:$PATH make -C ${IMAGE_VERSION} ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- modules_install INSTALL_MOD_PATH=$PWD/p2/
 PATH=$PWD/gcc/bin:$PATH make -C ${IMAGE_VERSION} ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- firmware_install INSTALL_FW_PATH=$PWD/p2/
+
+# Mali Kernel driver
+git clone https://github.com/superna9999/meson_gx_mali_450 -b DX910-SW-99002-r7p0-00rel1_meson_gx --depth 1
+(cd meson_gx_mali_450 && KDIR=$PWD/../$IMAGE_VERSION ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- ./build.sh)
+VER=$(ls p2/lib/modules/)
+sudo cp meson_gx_mali_450/mali.ko p2/lib/modules/$VER/kernel/
+sudo depmod -b p2/ -a $VER
+rm -fr meson_gx_mali_450
+
 mkdir -p p2/etc/apt/apt.conf.d p2/etc/dpkg/dpkg.cfg.d
 echo "force-unsafe-io" > "p2/etc/dpkg/dpkg.cfg.d/dpkg-unsafe-io"
-http_proxy="http://127.0.0.1:3142" qemu-debootstrap --arch arm64 xenial p2
+if [ -n "$PROXY" ] ; then
+	http_proxy="$PROXY" qemu-debootstrap --arch arm64 xenial p2
+else
+	qemu-debootstrap --arch arm64 xenial p2
+fi
 tee p2/etc/apt/sources.list.d/ubuntu-ports.list <<EOF
-deb http://ports.ubuntu.com/ubuntu-ports/ trusty main universe multiverse restricted
-deb http://ports.ubuntu.com/ubuntu-ports/ trusty-updates main universe multiverse restricted
-deb http://ports.ubuntu.com/ubuntu-ports/ trusty-security main universe multiverse restricted
+deb http://ports.ubuntu.com/ubuntu-ports/ xenial universe multiverse restricted
+deb http://ports.ubuntu.com/ubuntu-ports/ xenial-updates main universe multiverse restricted
+deb http://ports.ubuntu.com/ubuntu-ports/ xenial-security main universe multiverse restricted
 EOF
 tee p2/etc/fstab <<EOF
 /dev/root	/	btrfs	defaults,compress=lzo,noatime,subvol=@ 0 1
 EOF
-tee "p2/etc/apt/apt.conf.d/30proxy" <<EOF
+if [ -n "$PROXY" ] ; then
+	tee "p2/etc/apt/apt.conf.d/30proxy" <<EOF
 Acquire::http::proxy "http://127.0.0.1:3142";
 EOF
+fi
+
+# libMali X11
+wget https://github.com/superna9999/meson_gx_mali_450/releases/download/for-4.12/buildroot_openlinux_kernel_3.14_wayland_20170630_mali.tar.gz
+tar xfz buildroot_openlinux_kernel_3.14_wayland_20170630_mali.tar.gz
+rm buildroot_openlinux_kernel_3.14_wayland_20170630_mali.tar.gz
+
+mkdir -p p2/usr/lib/mali
+cp buildroot_openlinux/buildroot/package/meson-mali/lib/arm64/r7p0/m450/libMali.so p2/usr/lib/mali/
+cd p2/usr/lib/mali
+ln -s libMali.so libGLESv2.so.2.0
+ln -s libMali.so libGLESv1_CM.so.1.1
+ln -s libMali.so libEGL.so.1.4
+ln -s libGLESv2.so.2.0 libGLESv2.so.2
+ln -s libGLESv1_CM.so.1.1 libGLESv1_CM.so.1
+ln -s libEGL.so.1.4 libEGL.so.1
+ln -s libGLESv2.so.2 libGLESv2.so
+ln -s libGLESv1_CM.so.1 libGLESv1_CM.so
+ln -s libEGL.so.1 libEGL.so
+cd -
+cp -ar buildroot_openlinux/buildroot/package/meson-mali/include/* p2/usr/include/
+echo /usr/lib/mali > p2/etc/ld.so.conf.d/mali.conf
+rm -fr buildroot_openlinux
+
 cp /usr/bin/qemu-aarch64-static p2/usr/bin/
 cp stage2.sh p2/root
 mount -o bind /dev p2/dev
@@ -89,8 +131,12 @@ chroot p2 /root/stage2.sh
 umount p2/dev/pts
 umount p2/dev
 rm p2/root/stage2.sh
-rm p2/etc/apt/apt.conf.d/30proxy
+if [ -n "$PROXY" ] ; then
+	rm p2/etc/apt/apt.conf.d/30proxy
+fi
 rm p2/etc/dpkg/dpkg.cfg.d/dpkg-unsafe-io
+
+cp binary-amlogic/boot.init p1/
 
 umount p2
 umount p1
@@ -99,9 +145,9 @@ dd if=binary-amlogic/u-boot.bin.sd.bin of="${IMAGE_LOOP_DEV}" conv=fsync bs=1 co
 dd if=binary-amlogic/u-boot.bin.sd.bin of="${IMAGE_LOOP_DEV}" conv=fsync bs=512 skip=1 seek=1
 
 losetup -d "${IMAGE_LOOP_DEV}"
+mv "${IMAGE_FOLDER}${IMAGE_FILE_NAME}" "${IMAGE_FILE_NAME}"
 if [ $RAM -ne 0 ]; then
-	mv "${IMAGE_FOLDER}${IMAGE_FILE_NAME}" "${IMAGE_FILE_NAME}"
 	umount "${IMAGE_FOLDER}"
-	rmdir "${IMAGE_FOLDER}"
 fi
+rmdir "${IMAGE_FOLDER}"
 rmdir p1 p2
